requirements.txt
bcrypt==4.1.2
Pillow==10.1.0
pygame==2.5.2
psutil==5.9.6 